refactor: rename executable to malstatback and update Makefile commands feat: implement ScrapCmd and version command for CLI application chore: update .gitignore to exclude malstatback binary fix: replace http.Get with utils.NetClient.Get to enable timeouts
