+++ .gitignore
@@ -1,8 +1,8 @@
 .env
-.vscode
 out
 bin
 dist/
+malstatback
 
 # Created by https://www.toptal.com/developers/gitignore/api/goland+all,go
 # Edit at https://www.toptal.com/developers/gitignore?templates=goland+all,go
