+++ Makefile
@@ -6,13 +6,9 @@ endif
 
 SHELL := /bin/sh
 
-# The name of the executable (default is current directory name)
-TARGET := scrapper
+# The name of the executable
+TARGET := malstatback
 
-# These will be provided to the target
-BUILD := `git rev-parse HEAD`
-
-BIN_DIR ?= dist
 CSV ?= malstat.csv
 DB ?= ${DATABASE}
 REPOSITORY ?= reidaa
@@ -24,34 +20,25 @@ DOCKERTAG ?= latest
 all: build
 
 $(TARGET):
-	goreleaser build --clean --single-target --snapshot
+	go build -o $@ main.go
 
 build: $(TARGET)
 	@true
 
 clean:
-	rm -rf $(BIN_DIR)
+	rm -f $(TARGET)
 
-install:
-	go install $(LDFLAGS)
+re: clean build
+.PHONY: re
 
-uninstall: clean
-	rm -f $$(which ${TARGET})
+run: build
+	./$(TARGET) scrap --top 100 --db $(DB)
 
-snapshot:
-	goreleaser release --clean --snapshot
+run-help: build
+	./$(TARGET) help
 
-run: install
-	$(TARGET) scrap --top 100 --db $(DB)
-
-run-help: install
-	$(TARGET) help
-
-run-version: install
-	$(TARGET) version
-
-run-serve: install
-	$(TARGET) serve
+run-serve: build
+	./$(TARGET) serve
 
 ansible:
 	ansible-playbook deployments/ansible/deploy.yml -vv 
@@ -60,23 +47,3 @@ deploy: build ansible clean
 
 lint: build
 	golangci-lint run
-
-format: requirements
-	goimports -l -w *.go pkg cmd
-	gofumpt -l -w *.go pkg cmd
-
-ci_check:
-	go mod tidy
-	test -z "$(git status --porcelain)"
-	test -z $(shell gofmt -l *.go pkg cmd) || echo "[WARN] Fix formatting issues with 'make format'"
-	golangci-lint run
-	go vet ./...
-
-docker-build:
-	docker build -f $(DOCKERFILE) -t $(REPOSITORY)/$(TARGET):$(DOCKERTAG) .
-
-docker-run: docker-build
-	docker run --rm $(REPOSITORY)/$(TARGET):$(DOCKERTAG) scrap --top 100 --db $(DB)
-
-docker-run-ghcr:
-	docker run --rm ghcr.io/reidaa/malstat-scrapper:latest scrap --top 100 --db $(DB)
