+++ cmd/serve.go
@@ -1,0 +1,13 @@
+package cmd
+
+import (
+	"github.com/urfave/cli"
+)
+
+func ServeCmd(ctx *cli.Context) error {
+	// var port = ctx.Int("port")
+	// if err := internal.Serve(ctx.Int("port")); err != nil {
+	// 	return err
+	// }
+	return nil
+}
