--- internal/serve.go
@@ -1,9 +1,0 @@
-package internal
-
-func server() error {
-	return nil
-}
-
-func Serve(port int) error {
-	return nil
-}
