+++ pkg/jikan/anime.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net/http"
 	"time"
 
 	"malstat/scrapper/pkg/utils"
@@ -21,10 +20,13 @@ func AnimeByID(id int) (*Anime, error) {
 
 	utils.Debug.Printf("GET %s", url)
 
-	response, err := http.Get(url)
+	// response, err := http.Get(url)
+	response, err := utils.NetClient.Get(url)
 	if err != nil {
 		return nil, err
 	}
+
+	defer response.Body.Close()
 
 	if response.StatusCode >= 300 {
 		return nil, errors.New(response.Status)
