+++ pkg/utils/utils.go
@@ -2,7 +2,13 @@ package utils
 
 import (
 	"log"
+	"net/http"
 	"os"
+	"time"
+)
+
+const (
+	timeout = 10
 )
 
 func FileExists(filename string) bool {
@@ -14,14 +20,18 @@ func FileExists(filename string) bool {
 	return !info.IsDir()
 }
 
-// Info writes logs in the color blue with "INFO: " as prefix
+// Info writes logs in the color blue with "INFO: " as prefix.
 var Info = log.New(os.Stdout, "\u001b[34mINFO: \u001B[0m", log.LstdFlags)
 
-// Warning writes logs in the color yellow with "WARNING: " as prefix
+// Warning writes logs in the color yellow with "WARNING: " as prefix.
 var Warning = log.New(os.Stdout, "\u001b[33mWARNING: \u001B[0m", log.LstdFlags|log.Lshortfile)
 
-// Error writes logs in the color red with "ERROR: " as prefix
+// Error writes logs in the color red with "ERROR: " as prefix.
 var Error = log.New(os.Stdout, "\u001b[31mERROR: \u001b[0m", log.LstdFlags|log.Lshortfile)
 
-// Debug writes logs in the color cyan with "DEBUG: " as prefix
+// Debug writes logs in the color cyan with "DEBUG: " as prefix.
 var Debug = log.New(os.Stdout, "\u001b[36mDEBUG: \u001B[0m", log.LstdFlags|log.Lshortfile)
+
+var NetClient = &http.Client{
+	Timeout: time.Second * timeout,
+}
